requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.14.0
drf-nested-routers==0.94.1
Faker==25.2.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
